BE/requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.6.2
chardet==3.0.4  # Not needed for Python 3.6+, you can remove it if you're using a modern version
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
deep-translator==1.11.4
et-xmlfile==1.1.0
filelock==3.17.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.2.0
googletrans==4.0.0rc1
greenlet==3.1.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.28.1
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==2.8.7
nltk==3.9.1
numpy==1.24.0  # Ensure numpy version is compatible with your OpenCV and other libraries
openpyxl==3.1.4
packaging==24.2
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy>=1.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.37
sympy==1.13.1
textblob==0.19.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.1.3
xlrd==2.0.1
python-dotenv==0.21.0
flask-migrate==3.1.0
flask-cors==5.0.0
opencv-python-headless==4.8.0.76
pillow==9.5.0
